job-scraper-backend/requirements.txt
astroid==3.2.2
async-timeout==4.0.3
attrs==24.2.0
autopep8==2.3.1
beautifulsoup4==4.12.3
bleak==0.22.2
blinker==1.8.2
cachelib==0.9.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dill==0.3.8
Flask==3.0.3
Flask-Caching==2.3.0
flatbuffers==24.3.25
fonttools==4.53.1
frozendict==2.4.6
html5lib==1.1
humanfriendly==10.0
idna==3.7
imageio==2.35.1
importlib_metadata==8.5.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mccabe==0.7.0
mpmath==1.3.0
multitasking==0.0.11
nest-asyncio==1.6.0
networkx==2.8.8
numba==0.60.0
numpy
onnxruntime==1.19.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
peewee==3.17.7
pillow==10.4.0
platformdirs==4.2.2
plotly==5.24.1
pooch==1.8.2
protobuf==5.28.0
pycodestyle==2.12.0
pylint==3.2.5
PyMatting==1.1.12
pynput==1.7.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
retrying==1.3.4
rpds-py==0.20.0
scikit-image==0.24.0
scipy==1.11.2
six==1.16.0
soupsieve==2.6
sympy==1.13.2
ta==0.11.0
tenacity==9.0.0
tifffile==2024.8.30
tomli==2.0.1
tomlkit==0.12.5
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.0.3
yfinance==0.2.46
zipp==3.20.1
gunicorn
flask_cors
requests
python-dotenv
